Don't include articles in comments
The default messagebox behavior is to include article text when an
article is selected. For this extension, the article should go in
the main content area, but not in the internal comment box. This
commit passes a flag for MessageBox to not include the article text
in the comment box.

This fix requires a change in RT. A patch is included to apply the change.